requirements.txt
python>=3.6
pandas
openpyxl
python-docx